requirements.txt
BTrees==4.1.2
CourseCombine==0.0
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
PasteDeploy==1.5.2
Pygments==2.0.2
WTForms==2.0.2
WebOb==1.4.1
WebTest==2.0.18
ZConfig==3.0.4
ZEO==4.1.0
ZODB==4.1.0
ZODB3==3.11.0
argparse==1.2.1
beautifulsoup4==4.3.2
camtasia-request==0.0
gevent==1.0.2
greenlet==0.4.7
gunicorn==19.3.0
iso8601==0.1.10
mock==1.0.1
nose==1.3.7
peppercorn==0.5
persistent==4.0.9
pyramid==1.5.7
pyramid-debugtoolbar==2.3
pyramid-jinja2==2.5
pyramid-mailer==0.14.1
pyramid-mako==1.0.2
pyramid-tm==0.11
pyramid-wtforms==2.0.1
pyramid-zodbconn==0.7
repoze.lru==0.6
repoze.sendmail==4.2
requests==2.7.0
six==1.9.0
transaction==1.4.3
translationstring==1.3
venusian==1.0
waitress==0.8.9
wsgiref==0.1.2
zc.lockfile==1.1.0
zdaemon==4.1.0
zodburi==2.0
zope.deprecation==4.1.2
zope.interface==4.1.2
